requirements.txt
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
dell-recovery==0.0.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.2.0
Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.3
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.1.0
drf-yasg==1.21.7
duplicity==0.8.21
fasteners==0.14.1
future==0.18.2
gunicorn==21.2.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.1.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.1.0
progressbar==2.5
protobuf==3.12.4
psycopg2==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
sos==4.5.6
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
